Move SmartStreamModule from filter into smart_stream/mod
